services/google-mcp-sidecar/requirements.txt
google-api-python-client~=2.190.0
google-auth~=2.48.0
google-auth-oauthlib~=1.2.4
mcp~=1.26.0
uvicorn~=0.41.0
